+++ cmd/paramount/main.go
@@ -10,8 +10,8 @@ func main() {
    var guid string
    flag.StringVar(&guid, "b", "", "GUID")
    // f
-   var bandwidth int64
-   flag.Int64Var(&bandwidth, "f", 3_063_000, "target bandwidth")
+   var bandwidth int
+   flag.IntVar(&bandwidth, "f", 3_063_000, "target bandwidth")
    // i
    var info bool
    flag.BoolVar(&info, "i", false, "info")
